# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Revendis API",
    "version": "0.2.0",
    "description": "Initial OpenAPI contract for Sprint 1 endpoints.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.revendis.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/finance/receivables": {
      "get": {
        "summary": "List receivables",
        "operationId": "listReceivables",
        "tags": [
          "finance"
        ],
        "responses": {
          "200": {
            "description": "Receivables list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "summary": "Create receivable",
        "operationId": "createReceivable",
        "tags": [
          "finance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceivableInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Receivable created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/finance/receivables/{id}/settle": {
      "post": {
        "summary": "Settle receivable",
        "operationId": "settleReceivable",
        "tags": [
          "finance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceivableSettleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Receivable settled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
      "ReceivableInput": {
        "type": "object",
        "required": [
          "saleId",
          "amount",
          "dueDate"
        ],
        "properties": {
          "saleId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "method": {
            "type": "string"
          }
        }
      },
      "ReceivableSettleInput": {
        "type": "object",
        "required": [
          "amount",
          "settledAt"
        ],
        "properties": {
          "amount": {
            "type": "number"
          },
          "settledAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}